requirements.txt
openpyxl
pandas==2.0.1
plotly==5.13.1
streamlit==1.12.0
Pillow==9.2.0